refactor: moving to test utils
